finish chenges slider
