+++ src/js/container/slider.js
@@ -1,15 +1,17 @@
+import {images} from "../store.js"
+
 const main=document.getElementById('main')
 
 const slayder= document.createElement('div')
 slayder.classList.add("adaptivn-slayder")
 main.appendChild(slayder)
 
-const prev = document.createElement('a')
+const prev = document.createElement('button')
 prev.classList.add('prev')
 prev.innerText="❮"
 slayder.appendChild(prev)
 
-const next = document.createElement('a')
+const next = document.createElement('button')
 next.classList.add('next')
 next.innerText="❯"
 slayder.appendChild(next)
@@ -17,14 +19,6 @@ slayder.appendChild(next)
 const kadr= document.createElement("div")
 kadr.classList.add("kadr")
 main.appendChild(kadr)
-
-const images = [
-    'https://images.wbstatic.net/bners1/big_new__22.jpg',
-    'https://images.wbstatic.net/bners1/big_brand_17_12_22.jpg',
-    'https://images.wbstatic.net/bners1/big_cosmetics1.jpg',
-    'https://images.wbstatic.net/bners1/big_toys.jpg',
-    'https://images.wbstatic.net/bners1/big_inditexx_14_11_22.jpg',
-];
 
 images.forEach(src => {
     const img = document.createElement('img');
@@ -39,60 +33,63 @@ images.forEach(src => {
 
 const span =[1,2,3,4,5];
 
-span.forEach(src => {
-
-    const spans = document.createElement('span');
-    spans.className = 'point-kadr';
-    kadr.appendChild(spans)
-
-});
-
+for (let i = 0; i < span.length; i++) {
+  const spans = document.createElement('span');
+  spans.className = 'point-kadr';
+  kadr.appendChild(spans)
+  spans.addEventListener("click", (n) => {
+    slideIndex = i + 1
+    showSlides(slideIndex);
+  })
+}
 let timer = 0;
 makeTimer(); 
 function makeTimer(){
    clearInterval(timer) 
-   timer = setInterval(function(){
+   timer = setInterval(function (){
      slideIndex++;
      showSlides(slideIndex);
-   },2000);
+   },3000);
  }
 
  let slideIndex = 1;
  showSlides(slideIndex);
 
-next.addEventListener('onclick', function () {
+next.addEventListener('click', function () {
   showSlides(slideIndex += 1);
   makeTimer();
 })
 
-prev.addEventListener('onclick',function(){
+prev.addEventListener('click',function(){
     showSlides(slideIndex -= 1);
     makeTimer();  
 })
 
-// span.addEventListener('onclick',function(n) {
-//   showSlides(slideIndex = n);
-//   makeTimer();
-// })
-
-  
 function showSlides(n) {
   let i;
   let slides = document.getElementsByClassName("parag");
   let dots = document.getElementsByClassName("point-kadr");
   if (n > slides.length) {
-  slideIndex = 1
+    slideIndex = 1
   }
   if (n < 1) {
-  slideIndex = slides.length
+    slideIndex = slides.length
   }
   for (i = 0; i < slides.length; i++) {
-  slides[i].style.display = "none";
+    slides[i].style.display = "none";
   }
   for (i = 0; i < dots.length; i++) {
-  dots[i].className = dots[i].className.replace("deystvuyus", "");
+    dots[i].className = dots[i].className.replace("deystvuyus", "");
   }
-  slides[slideIndex - 1].style.display = "block";
-  dots[slideIndex - 1].className += " deystvuyus";
-}
 
+  slides[slideIndex - 1].style.display = "block";
+  dots[slideIndex - 1].classList.add("deystvuyus"); 
+}
+let hover = document.querySelector(".adaptivn-slayder");
+
+hover.addEventListener("mouseover",function(){
+  window.clearInterval(timer);
+}); 
+hover.addEventListener("mouseout",function(){
+  makeTimer()
+});
