+++ src/js/store.js
@@ -1,6 +1,6 @@
 //Здесь будут храниться наши данные, которые мы будем использовать для динамического создания карточек, товаров корзины и тд.
 
-export { cardData, basketData, sliderData }
+export { cardData, basketData, images}
 
 const cardData = [
     {
@@ -28,4 +28,10 @@ const cardData = [
 
 const basketData = []
 
-const sliderData = []
+const images = [
+    'https://images.wbstatic.net/bners1/big_new__22.jpg',
+    'https://images.wbstatic.net/bners1/big_brand_17_12_22.jpg',
+    'https://images.wbstatic.net/bners1/big_cosmetics1.jpg',
+    'https://images.wbstatic.net/bners1/big_toys.jpg',
+    'https://images.wbstatic.net/bners1/big_inditexx_14_11_22.jpg',
+]
